+++ activerecord/lib/active_record/connection_adapters/abstract/schema_statements.rb
@@ -1227,15 +1227,23 @@ module ActiveRecord
         options
       end
 
-      # Removes the given check constraint from the table.
+      # Removes the given check constraint from the table. Removing a check constraint
+      # that does not exist will raise an error.
       #
       #   remove_check_constraint :products, name: "price_check"
+      #
+      # To silently ignore a non-existent check constraint rather than raise an error,
+      # use the `if_exists` option.
+      #
+      #   remove_check_constraint :products, name: "price_check", if_exists: true
       #
       # The +expression+ parameter will be ignored if present. It can be helpful
       # to provide this in a migration's +change+ method so it can be reverted.
       # In that case, +expression+ will be used by #add_check_constraint.
       def remove_check_constraint(table_name, expression = nil, **options)
         return unless supports_check_constraints?
+
+        return if options[:if_exists] && !check_constraint_exists?(table_name, **options)
 
         chk_name_to_delete = check_constraint_for!(table_name, expression: expression, **options).name
 
@@ -1437,6 +1445,10 @@ module ActiveRecord
       end
 
       private
+	def check_constraint_exists?(table_name, **options)
+	  check_constraint_for(table_name, **options).present?
+	end
+
         def validate_change_column_null_argument!(value)
           unless value == true || value == false
             raise ArgumentError, "change_column_null expects a boolean value (true for NULL, false for NOT NULL). Got: #{value.inspect}"